// src/modules/whales/entities/tracking-address.interfaces.ts
import { type ChainKey } from '../../../common/interfaces/chain-key.interfaces';

export interface INormalizedAddressCandidate {
  readonly chainKey: ChainKey;
  readonly address: string;
}

export interface IResolvedTrackedWalletSubscription {
  readonly walletId: number;
  readonly chainKey: ChainKey;
  readonly walletAddress: string;
  readonly walletLabel: string | null;
}

export interface IResolvedHistoryTarget {
  readonly chainKey: ChainKey;
  readonly address: string;
  readonly walletId: number | null;
}